MB-31850: Introduce clusterCapabilities to terse bucket info ...
... pools/default/nodeServices API and also in CCCP payload.

clusterCapabilities as the name suggests will be the capabilities of the
cluster , i.e., what the server support, such as protocols, features,
security, etc. It is not intended that clusterCapabilities provide an
exhausting list of service features. Just those that are notable for
upgrade, and some important advertisable features, for example, some
dynamic features. It will not serve as a single source of truth.

There is a strict relationship between cluster_compat_mode and
capabilities, for example, only when cluster is 5.5.0 or higher enhanced
prepared statements for n1ql are supported.

Adding 2 new JSON fields, “clusterCapabilities” and
“clusterCapabilitiesVersion”, to pools/default/nodeServices and also
pools/default/b/<bucketname> REST API. This information will also be
present in the CCCP payload alongside bucketCapabilities, and will be
duplicated per bucket.
- clusterCapabilities: Has a per-service format.
- clusterCapabilitiesVersion: A version field([Major, Minor]) for
  capabilities, signifying the version number for the format of all
  per-service capabilities. That is, the version number will only bump
  up the version in case of format change of capabilities field, which
  is expected to be rare. Addition of capabilities within the service
  do not constitute a version increase.

"clusterCapabilitiesVersion" : [1, 0],
“clusterCapabilities”: {
    "kv" : ["kv_capability_example_1, kv_capability_example_2"],
    "n1ql" : ["enhanced_prepared_statements"]
}

Change-Id: Iaaa0e7a07ac5f74d0f81194b5b2695f26f1d6c28
Reviewed-on: http://review.couchbase.org/103700
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>