MB-32776: Introduce rebalance ID.
Rebalance and any automatic retries upon failure will be tracked
by a new unique ID.
The ID will be generated for each rebalance operation that is
manually triggered by the user.
If rebalance fails and is automatically retried, it will use the
same ID as the failed one.

The ID will be displayed in various user log messages for
rebalance and other operations that use rebalance code such as
graceful failover.

Note this rebalance ID is different from the one that is
currently passed by ns_server to the services.
ns_server generates a new ID before performing rebalance of each
topology aware service. A different ID is
generated for each service. That behaviour has not changed.

Change-Id: I924b3daac3851a9ed9ea83c981cd4c53f6f79ec4
Reviewed-on: http://review.couchbase.org/103980
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Poonam Dhavale <poonam@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>