Always terminate dependent async-s with reason 'shutdown'.
If we're explicitly terminating dependent async-s due to some kind of
error, it means that this error will be visible either because it's
propagated to the user of the async or because a crash report will be
generated by our async. Since the error is already visible, there's no
point in propagating it further down the async tree.

Change-Id: I73b48152a1e04db99e2e64e27d2f259f8f7eac64
Reviewed-on: http://review.couchbase.org/104297
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>