MB-32880: Apply config with topology on every janitor run.
Make sure we apply the topology information along with the state
information through the set_vbucket command to memcached during
janitor runs to keep information in ns_server and memcached in sync.

No attempt is made to,
1. Remove undefined from the map during failover.
2. Handle a case where increase in replica count adds an undefined in
the map.

Expectations from KV: KV engine should understand the undefined set in
the topology and ignore them.

Change-Id: I35eca77d036da4da0e8c0de0aded28a52da6be8c
Reviewed-on: http://review.couchbase.org/104233
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>