MB-32256: REST API to set/get cluster encryption...
...levels.

There are 2 levels in which the cluster encrytion can be configured
in: 'control' and 'all'.

'control' is the default value when cluster encryption is enabled.
At this level, only the control plane (the Erlang distribution) is
encrypted.

When the level is set to 'all', even the data channels will be
encrypted. This would include all the cross node communication
channels established by every service in Couchbase Server.

NOTE: Currently, cluster encryption level 'control' is a no-op. This
      means that the erlang distribution doesn't operate in encrypted
      mode when set to this level. However, this would be the case
      once we figure out the scheme to be used to handle the erlang
      distribution configuration.

Change-Id: Ie0c66daf784f8113f94bf521249f377817460096
Reviewed-on: http://review.couchbase.org/102827
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>