MB-32880: Do not change length of vbucket chain in ...
... ns_janitor when cluster > madhatter.

Pre mad-hatter, there is no change to the janitor behaviour.

Post mad-hatter changes that follow as a result of the above mentioned
change is,

1. Setting the vbucket map is allowed with different length chains.

2. Defer changing vbucket map to rebalance when changing num_replicas
i.e., when the admin changes number of replicas for the bucket, do not
update the vbucket map until rebalance is performed and not in the
janitor prior to rebalance.

3. Explicitly disallow delta-recovery for a bucket if vbucket map of
that bucket doesn't contain number of replicas according to Bucket
Config.

4. This changeset can result in varied length chains if rebalance was
interrupted for any reason. The janitor does not fix the length of the
chains.

Change-Id: Id7b0c50246796233de4ba688cfd5eaa81a53cad5
Reviewed-on: http://review.couchbase.org/105359
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>