Make adding new stages easy.
Adding a new stage is now equivalent to signalling the start of the
stage,
master_activity_events:note_rebalance_stage_started(Stage,
                                                    NodesInvolved)
If no nodes are involved in the stage i.e., NodesInvolved are [], we
ignore the stage as a part of rebalance visibility.

This can result in new stages showing up in the UI which weren't
part of the rebalance visibility output at the start of rebalance.

To mark stage as completed,
master_activity_events:note_rebalance_stage_completed(Stage)

Part of EPIC,
MB-30894: Rebalance visibility and reporting

Change-Id: I95f7542f4fa6b1e0771e2ab83879efb98ac48e03
Reviewed-on: http://review.couchbase.org/105567
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>