Simplify fun janitor_agent:apply_new_bucket_config/6.
The complexity of this function arose when we switched from chain
(A -> B -> C) to star (A -> B, A -> C) replication and we wanted the
transition to be as smooth as possible. The problem though was the old
nodes, we needed to coerce them into creating star replications
without changing their code. Hence the fake vbucket maps that this
function creates. Now that those times are long gone, this function
can pass untouched bucket config to the janitor_agent process. This
also removes the need for apply_new_config_replica_phase handling in
the janitor process.

Change-Id: I53a689f638243dd0f8a7cf9c5ce926e12bbb2fc8
Reviewed-on: http://review.couchbase.org/106733
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>