MB-31855: Goport to return exit code as 0 while...
...shutting down services.

Currently when ns_server asks the services to be shutdown (typically
during a failover), goport sends a SIGKILL to the actual service. That
exit code (137) gets propagated back to ns_server. This makes ns_server
log an error message to the user log and this can confuse the user.

Since ns_server is instructing the goport to shutdown the service, the
fix would be to change goport to return back an exit code of 0 when asked
to shutdown. Note that the case where the service actually dies before the
SIGKILL from goport could be processed gets masked but this behavior exists
even today.

However, goport will continue to return the proper exit code when the service
gets shutdown unexpectedly.

Change-Id: Ie5b7ebe7f54121a8bacce43f104c57741d3e735a
Reviewed-on: http://review.couchbase.org/111770
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>