MB-32256: do not allow to modify net config if autofailover is on
While changing the distribution protocol for a node in the cluster,
it's possible that node might appear to be down and result in an
auto-failover.

As we expect the administrator to turn on cluster encryption or
move from IPv4 to IPv6 in a controlled environment
(during their maintenance schedules) we have enforced the
auto-failover feature to be turned off. This would save
unnecessary rebalance operation required after the failover.

Change-Id: Id8ff645642e94bffd803774976cedf86857f9fc8
Reviewed-on: http://review.couchbase.org/107620
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>