MB-28525: non trivial map fixup for chains with failed over master.
For the first version we ignore FFMap and always select Map chain.
It is the promise of the rebalance process, that the old chain will
always have all durable mutations.

It is not possible to have a succesfull durable write on the new master
until topology is set and replications are established, which happens
after the map is updated with the new chain. Rebalance will wait until
the map update is propagated to all nodes.

It is possible that janitor will later on replace the old chain with
the single element chain [NewMaster]. This is not a best outcome, since
we lose the replicas, but we still guarantee that we won't lose durable
mutations, which at this moment are fully replicated to NewMaster.

Promote max replica i.e, make replica that is most ahead as active

A replica is considered ahead of another replica if its last snapshot
seqno is greater, if they are the same, the replica with greater
high_seqno is then considered ahead.

Change-Id: I66f51051815b07c276172732279742591e89dd7e
Reviewed-on: http://review.couchbase.org/109875
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>