MB-33749: Designate the same port number to be...
...used to host either an SSL or a non-SSL endpoint for projector.

As the projector process has been designed to re-use the same port
to either host a non-SSL or SSL endpoint, ns_server was using the
single port definition to also build projector's https command line
arguments.

In community edition, however, this logic would end up passing in the
certificate details on the command line to the projector process.
We shouldn't be passing in the certificate details in CE mode.

This patch defines separate port definitions (one for non-SSL and one
for SSL endpoint) for the projector process with both of them using
the same port number and uses the SSL port definition to build the
'https' command line arguments. This would allow the existing logic
in ns_server to not pass on the https arguments in CE mode.

Change-Id: Ifaea0f3edf0c6c6b26b6f8acf62c697d1a589e25
Reviewed-on: http://review.couchbase.org/107931
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>