MB-33828: On unexpected active node in Topology, rectify chain
with only active node.

In the event of node rename, which can occur if we are the only node in
the cluster, we will have updated the config and thereby the vbucket
map.  However, the topology in memcached will not be set, with new node
name.

If we discover that in the Topology in memcached of the active node,
doesn't point to the active node itself as the master, set chain with
active node as the master and maintaining the replicas.

Change-Id: Ifcfd4a028db552a729919f28b9331ef5f47febff
Reviewed-on: http://review.couchbase.org/107994
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>