MB-33750 Try working around the bug upon taking over mastership.
The theory is that many customers will see the issue in question when
upgrading. The fact they didn't encounter it before may mean two
things:

 - No master transitions happened in the observable past.
 - There were master transitions, but they got lucky.

The first scenario is a problematic one. The leader_lease_acquirer
workers have probably accumulated enough stale PIDs to trigger the
problem once the new master takes over.

This patch is a best effort attempt to prevent the upgrade from
triggering the issue. When the fixed node takes over mastership, it
will attempt to kill the leader_lease_acquirer process on the old
master. This may or may not kill all of leader_lease_acquirer's worker
processes: those that don't die should get stuck in the shutdown. But
at least they are not going to be affecting the operation of the new
master anymore.

Change-Id: Ia812cce034a5708014f3e4c0415c24628e5cf423
Reviewed-on: http://review.couchbase.org/108003
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>