MB-32256: Apply generated CA on joinee when adding the node
so if the dist is set up to be encrypted, the joinee can establish
initial tls connection to the cluster and pull the config. In this case
temporary the joinee will have node certificate that is out of
sync with CA. The situation will be fixed as soon as the config
is pulled bringing the CA + private key to the joinee.

Change-Id: Icdb410b56524482103274a1fed619a2a96e61ddd
Reviewed-on: http://review.couchbase.org/108289
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>