Disable dual topology before updating the replications.
In case the vbucket move was not a takeover move, there'll be already
replications to the new replicas prior to
update_replications_post_move() call. So not updating the topology
before the call unnecessarily makes sync writes unavailable for the
duration of update_replications_post_move() (since it removes the
replications to the old replicas).

For takeover moves, sync-writes are unavailable for the duration of
update_replication_post_move() regardless of when the topology is
updated. But since update_replication_post_move() doesn't only create
replications to the new replicas, but also removes old replications,
this patch potentially still has a positive impact.

Change-Id: Ib81a52a2c4c25f53843c9bb1382c273f7277a0e6
Reviewed-on: http://review.couchbase.org/108524
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>