Make goport args readable when logged.
Currently, some arguments passed into the some services are not
logged as strings but as a list of ASCII values in the babysitter log.
This makes it difficult to figure out what was actually passed when
looking at the logs. This fix prints them all as strings in the log
file.

Change-Id: Idd20a4d0fba2aba8dc5ee01f9ec1c6612d046139
Reviewed-on: http://review.couchbase.org/108608
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>