MB-33781 Return correct CBASDirs
Prior to this change the CBASDirs returned defaulted to the one on which
the REST Api was run.  e.g. if the API came in on node2 then node2's
default CBASDir was used even for the other nodes in the cluster.

This change returns the per-node CBASDirs value which is initialized
to the default path in the config during init if it hasn't been set
by the user.

Change-Id: Ic72ce6531f430150612991468ee502d042add0eb
Reviewed-on: http://review.couchbase.org/108677
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Reviewed-by: Ajit Yagaty <ajit.yagaty@couchbase.com>