Simplify ns_memcached:do_terminate/4 further.
This changes what messages are logged to the user log. Now they
directly correspond to what's going to happen to the bucket in
memcached. If the bucket is going to get deleted, we log the "Shutting
down bucket" message. If the bucket is not getting deleted, we log
"Control connection disconnected" message. This makes the nested case
unnecessary.

Change-Id: I31404d3c6560e4d74a07e8cbdc6d37a7c5291266
Reviewed-on: http://review.couchbase.org/108930
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>