Don't allow to disable nodeEncryption when...
...the cluster encryption level has been set to 'all'.

With cluster encryption level set to 'all', the services will be
interacting using SSL connections. We shouldn't allow the encryption
feature to be turned off at a node level when all other services are
using the encrypted connections. The administrator will have to first
set the cluster encryption level to 'control' and then attempt to
disable encryption on a node.

Change-Id: If1a4c65365a77d394897013c91e9c2f37df8fa20
Reviewed-on: http://review.couchbase.org/109152
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>