MB-32256: MB-34278: Migration to MadHatter for ns_server nodename
... to prevent node from removing itself from cluster after
offline upgrade

If there is no "ip" file, it might be the case that
node has default name. The problem is that default
pre-MH name and default MH name are different. Hence the
migration. The idea is the following: if "nodefile"
exists, but there is no ip or ip_start files, we should
use 127.0.0.1 (or ::1) address (instead of cb.local)

Change-Id: Ic5b7418f114849c9b3be40bc8fe6eef17f812cf8
Reviewed-on: http://review.couchbase.org/109352
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>