Prevent convergence of triq_utils:random_integer_fun/0.
Explicitly limit number of recursive applications of base functions to
a low number. Otherwise, the resulting functions tend to converge to
the same result irrespective of the input. That is because whenever
there's a const() in the sequence, the end result is going to be the
same for any input. Similarly the convergence can be caused by integer
division producing same results for different inputs.

Change-Id: I707ad2e3caa1631312e3a5940c92a940de1ca712
Reviewed-on: http://review.couchbase.org/109749
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>