Extend ns_bucket_worker to support "transient" buckets.
Where "transient" means that the bucket does not exist on this node
according to the config, but we want to start the bucket related
processes anyway.

This is going to be used for reorchestration of delta recovery where
we want to have some preparation steps on a node before making it part
of the bucket.

For simplicity, currently only one process at any given moment of time
is allowed to keep transient buckets. Additionally, all of them need
to be started together, there's no way to start additional transient
buckets.

Change-Id: I9c11ad8b3883bd11e1de0dcc158033e8204845cb
Reviewed-on: http://review.couchbase.org/109756
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>