Don't allow cluster encryption level to be set to all when...
...client certificate authentication state is mandatory.

With client certificate authentication state set to 'mandatory', the
client must provide a certificate for authentication when connecting to
a TLS server. However, this will be a problem when cluster encryption
level gets set to 'all' as the services will have to supply a client
certificate, which we don't support, when trying to connect to the
servers of interest for intra-cluster communication. This fix prevents
the user from getting into such a state.

Change-Id: Id2f889dbf8e78b9f7dcaa5b9efcf5043b18958f9
Reviewed-on: http://review.couchbase.org/109601
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ajit Yagaty <ajit.yagaty@couchbase.com>