Stop consulting topologies in ns_janitor.
The initial motivation for storing topologies persistently in
memcached was to work around the weak consistency model of
ns_config. Unfortunately we've encountered couple of issues with this
approach:

 1. While handling failover we don't have access to the node that's
    being failed over to consult its topologies. Obvious in hindsight,
    but was not obvious when the decision was made.

 2. Even in the cases where we can consult the topology information
    from memcached, it's complicated by the fact that ep-engine
    persists this information asynchronously and there's no way for
    ns_server to easily know when it's safe to assume that the
    topology won't rollback to some previous state.

So we decided to "super sync" the ns_config instead to get the
consistency guarantees we need (or to get as close as possible to that
goal). This will be done in following commits.

This patch partially reverts the following commits:

  d81d0ad490e923d7726ae2e8111bac8f377be534
  fe73eb5242edf46c6376b9ffb142c40efa479e96
  5839c1cb2df93ed6cb14a158342b865caaf7148d
  9772c190eba7c777391622f2d316a6f9779e19a8

Change-Id: I117a2b9d060bdde527b41f97a212d9bc25187db1
Reviewed-on: http://review.couchbase.org/109937
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>