Update listeners and preferred protocols in cb_dist atomically
When node is being added to a cluster it needs to apply cluster's
listeners and cluster's preferred protocols. It might not always
be possible to change those values separatelly, since the listeners
of the cluster might not be valid when applied to current joinee's
configuration. In other words, transitional state will be invalid.

For example, cluster has the following config:
listeners: inet6_tls
preferred_proto: inet6_tls

when joinee's config is
listeners: inet_tcp
preffered_proto: inet_tcp

In this configuration joinee can't apply listeners first because it
will have to disable inet_tcp listener in this case, which can't be
done, since it's being used.

In order to avoid this invalid transitional state, the whole config
should be applied atomically. By "atomic" here I mean a single change
of cb_dist configuration.

Change-Id: I86eb3acd18394fd6864ca45f622e9ca21e286900
Reviewed-on: http://review.couchbase.org/109947
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>