Make node rename synchronous
Rename of the node leads to generation of node's certs, which leads
to restart of web server. External components (UI and CLI) assume the
rename command to be synchronous. They can perform calls to web server
immediately after rename is done. In order to avoid the race we need
to make sure web server is restarted by the time the rename is finished.

Change-Id: I074855b934d6936c6d0c18dda5c28772d9016dc1
Reviewed-on: http://review.couchbase.org/110250
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>