Fix backfill accounting when index aware rebalance is disabled.
ns_vbucket_mover automatically accounts for the corresponding backfill
when it receives move_done message. To the contrary,
move_done_new_style doesn't do the same and an explicit backfill_done
needs to be sent prior to it. The original TAP rebalancer would not
send backfill_done when index aware rebalance was disabled, so
choosing either move_done or move_done_new_style depending on the
value of IndexAware somewhat made sense. But DCP rebalancer has always
sent backfill_done no matter what. The end result is that backfills
get accounted for twice with index aware rebalance disabled. And
therefore limits on the number of concurrent backfills are not
enforced.

Change-Id: I33a1a5c898f91f941802e711b9103e6758014f82
Reviewed-on: http://review.couchbase.org/110409
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>