MB-34081: Fix logic to return in_progress when a rebalance...
...operation is already underway and a new rebalance request is issued.

We have introduced a new field to be passed to the "start_rebalance"
message with the retry rebalance feature. It looks like we have missed
adding this new field in the function pattern matching logic for the
"rebalancing" state in ns_orchestrator. This results in ns_orchestrator
returning "rebalance_running" rather than "in_progress" which in turn
breaks behavior in the REST API handlers.

Change-Id: I2347f9e650e09c394e886da21f80bad592d05dfc
Reviewed-on: http://review.couchbase.org/110507
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>