Remove redundant checks in view compaction uninhibition.
Don't check if index aware rebalance is disabled or if view
compactions were asked to be ignored. It's enough to check for an
active inhibition in the ets table. If any of these options were used,
there'll be no inhibition in the ets table: the inhibition
code checks for these options as well.

Change-Id: Ic19fdb310f8928855e76060e2954004391a8619b
Reviewed-on: http://review.couchbase.org/111011
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>