Update vbucket map before the end of a move.
Previously we updated the vbucket map with the fast-forward chain for
a vbucket only after its move was fully completed. The issue with this
is that for a short time the vbucket map would refer to the old chain
that might have been partially or fully destroyed by the vbucket
mover. A failover of the orchestrator node could then lead to all data
being lost in the vbucket.

Consider the following. Node A is the orchestrator and there's a move
for some vbucket that looks like [A,B] -> [A,C]. The sequence of steps
for this move would be (ignoring the details that are irrelevant).

1. Do the move.
2. Set the sync-writes topology to [A,C].
3. Drop the old replication from A to B.
4. Delete the now unneeded vbucket copy on B.
5. Update the vbucket map with the fast-forward chain [A,C].

Now, let's imagine that the orchestrator node A goes down and gets
failed over in between the steps 4 and 5. Failover code believes that
the current chain for the vbucket is [A,B], so it fixes it up to
[B,undefined]. But there's no copy of this vbucket on B anymore, so
the data is lost.

For a slightly different scenario, let's imagine that the orchestrator
fails somewhere between the steps 2 and 4. In this case node B will
still have the vbucket copy, so we won't lose all data in this
vbucket, but there might have been sync-writes acknowledged in the new
topology that will be missing from B.

This commit partially addresses this issue by making sure that the
vbucket map is updated before the step 2. In addition, this patch will
also synchronize ns_config to all other nodes, and won't proceed any
further if the synchronization fails.

Issues that are not currently handled:

 - ns_config_rep:ensure_config_seen_by_nodes() doesn't wait for the
   config to be persisted on destination nodes
 - ns_config_rep:ensure_config_seen_by_nodes() may fail to synchronize
   config without returning any errors
 - ns_janitor and failover don't do their part in config
   synchronization

Change-Id: Id7d1c26bd03298eaf9b6c2fd5d271050e3d6f34d
Reviewed-on: http://review.couchbase.org/111016
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>