Log config save failure only if we're not saving it again.
If the failure is persisting, we'll see it anyway. If it's
intermittent, maybe it's fine to ignore it.

The main motivation for the change is to restructure the code to be
more amenable for following changes. I'd actually prefer not to ignore
these failures at all. But we know that during node rename save config
might fail because we're trying to reach encryption_service running on
the babysitter. I don't want that to crash ns_config, but I'm also not
ready to address the underlying issue.

Change-Id: I9d0bd3ea9210ca151e30fc4678e0c053ff05e756
Reviewed-on: http://review.couchbase.org/111023
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>