Add ns_config:ensure_persisted/0.
That returns ok if all updates that happened before the call are
persisted to disk. The call itself does not initiate persistence, it
relies on the fact that updates initiate persistence themselves. So
the call simply waits for previously initiated saves to complete.

Note, the way this is currently implemented is prone to starvation:
ongoing updates will set pending_move_save to true, while
ensure_persisted() callers are replied to only once there are no more
pending persistence requests. I might address it later if this becomes
a problem.

Change-Id: Ic26a4b8c6888b8d9e8034a53081b3b02237887a8
Reviewed-on: http://review.couchbase.org/111024
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>