MB-32174 Change gotraceback on upgrade
On upgrade check if gotraceback is set to <<"crash">>.  If so change
the value to <<"single">>.  The reason is dumps obtained when "crash"
is specified are no longer needed and lead to a lengthy outage after
a crash to obtain them.

This does an online upgrade but doesn't restart any dependant service
to avoid what could be a lengthy outage just to set someting that is
only used when the service crashes.

Change-Id: I2f4050db5f6a845ebbc51088686d94070bdf2eb3
Reviewed-on: http://review.couchbase.org/110773
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Dave Finlay <dave.finlay@couchbase.com>