MB-34639: Handle not_running return from get_rebalance_info
Generally we don't fetch the rebalance info if rebalance is not running.
If however, rebalance is running when we check with the orchestrator but
has stopped by the time we get the info. We pass along the not_running
atom to the caller, which doesn't like it, and an exception is thrown
when we try to encode it to a json.

If the ns_rebalance_observer is not_running we expect the rebalance also
to terminate, which will be seen in the state of next refresh of taks
API.

Change-Id: I414c9210a47da63804146b3f15e54437e617ae81
Reviewed-on: http://review.couchbase.org/110973
Tested-by: Abhijeeth Nuthan <abhijeeth.nuthan@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ajit Yagaty <ajit.yagaty@couchbase.com>