Pull config before durability aware failover.
From all nodes that are not already failed over and are not being
failed over.

Even though rebalance and janitor will ensure that failover can always
ignore fast forward map (or in other words, using regular map will
never lose acknowledged durable writes), the ns_config may still cause
trouble in the scenario like the following.

 1 Rebalance is orchestrated by node A, a move from [A,B] to [A,C] is
   in progress.
 2 When it's safe to switch to the new topology, A updates the vbucket
   map and waits until the change is replicated to all other
   nodes. But A goes down before replicating the config update.
 3 Another node becomes an orchestrator and fails node A over.
 4 Node A comes back online. The unreplicated config update from 1 is
   now replicated to the rest of the cluster causing a conflict.

Normally such conflict would get resolved in favor of the new
orchestrator based on a later timestamp. But it's not impossible that
because of time mis-synchronization it will get resolved in favor of
the change produced in step 1. With the tools we have it's not easy
prevent this pathological situation. This change is an attempt to
lower the probability of this happening: if node A in step 1 manages
to replicate its update to at least one other node, the failover will
be able to pick it up.

Change-Id: Ie0d35d2be0d47e5323eee0cb0cc00d92f8475264
Reviewed-on: http://review.couchbase.org/111021
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>