Synchronize config in ns_janitor.
Whenever ns_janitor detects that the vbucket states it observes do not
match the vbucket map exactly, it synchronizes config to and from all
non-failed-over nodes. Pulling from is required to prevent config
update conflicts similar to the ones described in the previous
commit. Pushing config before changing any vbucket states is required
to guarantee the same invariant that the rebalancer maintains: that if
both regular and fast-forward chains are present, the regular chain
will always have all durable writes.

One thing to note regarding the set of nodes the config is
synchronized with: failed-over nodes are ignored, given that they most
likely are unhealthy. But it's also possible that they are healthy,
and the orchestrator may actually be running (or might have been
running not long ago) on one of these nodes. So in reality we need to
be synchronizing with these nodes too. But instead we invoke our "we
support one failure only" mantra: since there are failed-over nodes,
the limit on the number of supported failures is already exceeded.

Change-Id: Ia52c5c5a76b0467fee45f92dff9e82e3c840e506
Reviewed-on: http://review.couchbase.org/111022
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Aliaksey Artamonau <aliaksey.artamonau@couchbase.com>